+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Unreleased
 - Typography: Updated monospace font to Roboto Mono for consistency across login.gov sites. (#5891)
 - Multi-factor authentication: Add ability to opt numbers back in to receiving SMS that have replied "STOP" (#5894)
 - Icons: Replaced custom button icons using U.S. Web Design system icons. (#5904)
+- Analytics: Add ability to track when users go to Webauthn page to authenticate their device. (#5920)
 - User Interface: Updated styling for process lists following U.S. Web Design System best practices. (#5863)
 
 ### Accessibility